# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Natours API
  version: 1.0.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /booking/webhook:
    post:
      summary: Receive signed Stripe events
      responses:
        '200': { description: Event processed }
        '400': { description: Invalid signature }
...